pyproject.toml
[project]
name = "odoo-addons-openg2p-social-registry"
version = "0.0.0"
dependencies = [
    "odoo-addon-g2p_odk_user_mapping @ {root:uri}/g2p_odk_user_mapping",
    "odoo-addon-g2p_registry_g2p_connect_rest_api @ {root:uri}/g2p_registry_g2p_connect_rest_api",
    "odoo-addon-g2p_registry_id_deduplication @ {root:uri}/g2p_registry_id_deduplication",
    "odoo-addon-g2p_registry_deduplication_deduplicator @ {root:uri}/g2p_registry_deduplication_deduplicator",
    "odoo-addon-g2p_social_registry @ {root:uri}/g2p_social_registry",
    "odoo-addon-g2p_social_registry_theme @ {root:uri}/g2p_social_registry_theme",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
bypass-selection = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
